M4 grade-2: rebaseline edge subsystems to FreeBSD 15.0
spec 05 §2.4 lists bsm/ddb/netgraph/netpfil/netipsec as M4 cp targets.
Measured diff -rq vs 15.0 baseline before grade-2:
  * bsm:      8 differ
  * ddb:      29 differ + 2 only-15
  * netgraph: 1 differ
  * netpfil:  74 differ + 11 only-15 + 1 only-fs
  * netipsec: 0 differ (already updated in M2 / Phase 5b)

Rebaseline strategy:
  cp -af freebsd-src-releng-15.0/sys/<dir>/. f-stack/freebsd/<dir>/

Why a flat overwrite is safe here:
  * netgraph / netpfil / netipsec are gated by FF_NETGRAPH / FF_IPFW /
    FF_IPSEC and are NOT in the default lib SRCS (only ng_ipfw.c is
    referenced under FF_IPFW which is also disabled).
  * bsm / ddb are header-only or default-disabled.
  * No fstack-13 deltas were carrying user-visible behavior here.
  * The libfstack.a strict link (DP-M4-3=A) already passed before this
    commit, so nothing in the default build path depends on these
    legacy carry-over bits.

Note: ng_socket.c had a single-hunk fstack delta (#ifndef FSTACK around
the auto-load NETGRAPH module path). It is intentionally dropped in the
overwrite; FF_NETGRAPH stays disabled by default so this has no impact
on the M4 deliverable. It will be reapplied in M5 when FF_NETGRAPH is
enabled as part of optional-feature compile validation.

Result: all 5 subsystems show 0 differ + 0 only-15 vs 15.0 baseline.

+++ freebsd/netpfil/ipfw/nat64/nat64clat.h
@@ -1,5 +1,5 @@
 /*-
- * SPDX-License-Identifier: BSD-2-Clause-FreeBSD
+ * SPDX-License-Identifier: BSD-2-Clause
  *
  * Copyright (c) 2019 Yandex LLC
  * Copyright (c) 2019 Andrey V. Elsukov <ae@FreeBSD.org>
@@ -25,8 +25,6 @@
  * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * $FreeBSD$
  */
 
 #ifndef	_IP_FW_NAT64CLAT_H_
@@ -44,7 +42,7 @@ struct nat64clat_cfg {
 	char			name[64];
 };
 
-VNET_DECLARE(uint16_t, nat64clat_eid);
+VNET_DECLARE(uint32_t, nat64clat_eid);
 #define	V_nat64clat_eid	VNET(nat64clat_eid)
 #define	IPFW_TLV_NAT64CLAT_NAME	IPFW_TLV_EACTION_NAME(V_nat64clat_eid)
